Users queries are now imported into the main database file.
